+++ src/commands/projects/create/sql.ts
@@ -4,9 +4,9 @@ import type { CommandOptions } from '../../../utils/option-parser';
 import type * as OclifConfig from '@oclif/config';
 
 import AbstractProjectCreateCommand from '../../../abstract-project-create-command';
+import * as projectCreateOptions from '../../../services/projects/create/options';
 import Agents from '../../../utils/agents';
 import { optionsToFlags } from '../../../utils/option-parser';
-import * as projectCreateOptions from '../../../utils/options';
 
 export default class SqlCommand extends AbstractProjectCreateCommand {
   protected static options: CommandOptions = {
